Added deposit and withdraw. Now uses pointers.
The data structures containing the different sortings
of accounts now contains pointers to accounts and functions
have been adjusted to match this. The reason for this is that
updating an account in accounts_by_number needs  to affect the
same account in accounts_by_owner. This also means that accounts
are dynamically allocated and destroyed in the bank destructor.